move convert and compare to submodule "cli" now split em
